requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.3.2
pymongo==4.6.1
httpx==0.26.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
